+++ codegen/src/rhai_module.rs
@@ -162,6 +162,7 @@ pub(crate) fn generate_body(
             pub fn rhai_module_generate() -> Module {
                 let mut m = Module::new();
                 rhai_generate_into_module(&mut m, false);
+                m.build_index();
                 m
             }
             #[allow(unused_mut)]